+++ src/components/header.js
@@ -5,9 +5,6 @@ import {Images} from "../images/index"
 import Twitter from "../images/twitter.svg"
 import Github from "../images/github.svg"
 import { Button, Form, Container, Row, Col } from "react-bootstrap"
-import { AiOutlineSearch } from "react-icons/ai"
-import { MdChevronLeft } from "react-icons/md"
-import { IconContext } from "react-icons"
 import DgraphLogo from "../images/graphql-logo.png"
 import algoliasearch from "algoliasearch/lite"
 import {
@@ -19,7 +16,7 @@ import {
   connectStateResults
 } from "react-instantsearch-dom"
 import { CustomInstantSearch } from "./InstantSearchBar"
-
+import BackButtonMainWebsite from './MainWebsiteRedirect';
 
 const searchClient = algoliasearch(
   "0AG2ENC6EL",
@@ -71,50 +68,29 @@ const SearchBar = () => (
   </InstantSearch>
 )
 
-const BackIcon = () => {
-  return (
-    <>
-      <IconContext.Provider value={{ color: "#B7B7B7" }}>
-        <MdChevronLeft viewBox="0 0 12 20" />
-      </IconContext.Provider>
-      <MdChevronLeft viewBox="7 0 20 20" />
-    </>
-  )
-}
 
-const BackButtonMainWebsite = () => {
-  return (
-    <div className="back-button-main-website">
-      <Button
-        as="a"
-        href="https://dgraph.io/"
-        bsPrefix="navigate-main-website-button ml-auto"
-      >
-        <BackIcon />
-        Back to Main Website
-      </Button>
-    </div>
-  )
-}
 
 const Header = (props) => {
   return (
     <div className="header-search-main-site-container">
       <div className="d-flex justify-content-between border-bottom align-items-center">
-        <div className="page-logo d-md-flex d-lg-none">
+        <div className="page-logo d-xs-flex justify-content-between d-sm-flex d-md-flex d-lg-none">
           <div className="menu-icon-container-tablet d-inline-flex">
             <Button variant="light" bsPrefix="menu-button" onClick={()=>{
               props.showSideBar(!props.isSideBarVisible)
-            }}><img src={Images.menu_icon} alt="menuIcon" /></Button>
+            }}><img src={Images.menu_icon} alt="menuIcon"/></Button>
           </div>
           <Link to="/" className="img-logo header-link">
-            <img src={DgraphLogo} alt="Dgraph logo" />
+            <img src={DgraphLogo} alt="Dgraph logo" className="full-logo-button"/>
+            <img src={Images.partial_logo_icon} alt="menuIcon" className="partial-logo-button"/>
           </Link>
         </div>
       <div className="topbar d-flex page-header">
         <SearchBar />
         <div className="page-header justify-content-end">
+          <div className="back-to-mainwebsite-container">
           <BackButtonMainWebsite />
+          </div>
           <div className="header-icon-containers">
             <a
               href="https://twitter.com/dgraphlabs"
