+++ src/images/index.js
@@ -6,5 +6,6 @@ export const Images = {
   wrench: require("../images/icons/wrench.png"),
   clone: require("../images/icons/clone.png"),
   dropdown: require('../images/icons/Union (2).svg'),
-  menu_icon: require('../images/icons/menu_24px.png')
+  menu_icon: require('../images/icons/menu_24px.png'),
+  partial_logo_icon: require('../images/icons/partial_logo.svg')
 }
